+++ rusk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for wasm module shading feature
+
 ## [1.4.1] - 2025-12-04
 
 ### Added